move progress bar
